pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rbac-rag-mcp"
version = "1.0.0"
description = "RBAC-gated RAG MCP Server — intelligent company assistant with role-based data access"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9",
    "sentence-transformers>=2.2",
    "numpy>=1.24",
    "anthropic>=0.30",
    "fastmcp>=0.1",
    "python-dotenv>=1.0",
    "transformers>=4.40",
    "torch>=2.0",
    "accelerate>=0.27",
    "langchain>=0.2",
    "langchain-community>=0.2",
    "langchain-core>=0.2",
    "langchain-text-splitters>=0.2",
    "fastapi>=0.111",
    "uvicorn>=0.29",
    "python-multipart>=0.0.6",
    "pypdf>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
